requirements.txt
pandas
pdfminer.six
python-dotenv
OperationBattleshipCommonUtilities
fastapi
uvicorn[standard]
python-multipart
requests
requests
reportlab
boto3
markdown2
pdfkit